feat: introduce rslib to package the plugin and its declarations
