+++ .gitignore
@@ -37,3 +37,6 @@ yarn-error.log*
 
 # Dist folder
 dist
+
+# Profile
+.rslib/
