+++ rslib.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from "@rslib/core";
+
+export default defineConfig({
+  lib: [
+    {
+      format: "esm",
+      syntax: ["node 18"],
+      dts: {
+        bundle: true,
+      },
+    },
+  ],
+});
